# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * News Card Generator Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  plan: string;
  apiKey: string;
  imagesToday: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface Template {
  id: number;
  userId: number;
  name: string;
  slug?: string | null;
  backgroundUrl?: string | null;
  logoUrl?: string | null;
  overlayUrl?: string | null;
  elements: string;
  category: string;
  aspectRatio: string;
  bannerColor: string;
  bannerGradient?: string | null;
  textColor: string;
  labelColor: string;
  font: string;
  fontSize: number;
  fontWeight: number;
  photoHeight: number;
  subtitle?: string | null;
  label?: string | null;
  logoText?: string | null;
  logoPos: string;
  logoInvert: boolean;
  textShadow: boolean;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTemplateBody {
  name: string;
  slug?: string | null;
  backgroundUrl?: string | null;
  logoUrl?: string | null;
  overlayUrl?: string | null;
  elements?: string;
  category?: string;
  aspectRatio?: string;
  bannerColor?: string;
  bannerGradient?: string | null;
  textColor?: string;
  labelColor?: string;
  font?: string;
  fontSize?: number;
  fontWeight?: number;
  photoHeight?: number;
  subtitle?: string | null;
  label?: string | null;
  logoText?: string | null;
  logoPos?: string;
  logoInvert?: boolean;
  textShadow?: boolean;
  isPublic?: boolean;
}

export interface GenerateBody {
  title: string;
  subtitle?: string | null;
  label?: string | null;
  templateId?: string | number | null;
  aspectRatio?: string;
  backgroundPhotoFilename?: string | null;
  logoPhotoFilename?: string | null;
}

export interface GeneratedImage {
  id: number;
  userId: number;
  templateId?: number | null;
  title: string;
  imageUrl: string;
  aspectRatio: string;
  fileSize?: number | null;
  createdAt: string;
}

export interface HistoryResponse {
  images: GeneratedImage[];
  total: number;
}

export interface ApiKeyResponse {
  apiKey: string;
}

export interface StatsResponse {
  totalImages: number;
  imagesToday: number;
  totalTemplates: number;
  plan: string;
  dailyLimit: number;
}

export type ListHistoryParams = {
  limit?: number;
  offset?: number;
};