# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: News Card Generator Pro API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: templates
    description: Card templates
  - name: generate
    description: Image generation
  - name: history
    description: Generation history
  - name: keys
    description: API key management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List user templates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"

  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /generate:
    post:
      operationId: generateImage
      tags: [generate]
      summary: Generate a news card image
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBody"
      responses:
        "201":
          description: Generated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: Get generation history
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"

  /keys/regenerate:
    post:
      operationId: regenerateApiKey
      tags: [keys]
      summary: Regenerate API key
      security:
        - bearerAuth: []
      responses:
        "200":
          description: New API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyResponse"

  /stats:
    get:
      operationId: getStats
      tags: [history]
      summary: Get dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        plan:
          type: string
        apiKey:
          type: string
        imagesToday:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, plan, apiKey, imagesToday, createdAt]

    Template:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        slug:
          type: string
          nullable: true
        backgroundUrl:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        overlayUrl:
          type: string
          nullable: true
        elements:
          type: string
        category:
          type: string
        aspectRatio:
          type: string
        bannerColor:
          type: string
        bannerGradient:
          type: string
          nullable: true
        textColor:
          type: string
        labelColor:
          type: string
        font:
          type: string
        fontSize:
          type: integer
        fontWeight:
          type: integer
        photoHeight:
          type: integer
        subtitle:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        logoText:
          type: string
          nullable: true
        logoPos:
          type: string
        logoInvert:
          type: boolean
        textShadow:
          type: boolean
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, name, elements, category, aspectRatio, bannerColor, textColor, labelColor, font, fontSize, fontWeight, photoHeight, logoPos, logoInvert, textShadow, isPublic, createdAt, updatedAt]

    CreateTemplateBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
          nullable: true
        backgroundUrl:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        overlayUrl:
          type: string
          nullable: true
        elements:
          type: string
        category:
          type: string
        aspectRatio:
          type: string
        bannerColor:
          type: string
        bannerGradient:
          type: string
          nullable: true
        textColor:
          type: string
        labelColor:
          type: string
        font:
          type: string
        fontSize:
          type: integer
        fontWeight:
          type: integer
        photoHeight:
          type: integer
        subtitle:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        logoText:
          type: string
          nullable: true
        logoPos:
          type: string
        logoInvert:
          type: boolean
        textShadow:
          type: boolean
        isPublic:
          type: boolean
      required: [name]

    GenerateBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        templateId:
          nullable: true
          oneOf:
            - type: string
            - type: integer
        aspectRatio:
          type: string
        backgroundPhotoFilename:
          type: string
          nullable: true
        logoPhotoFilename:
          type: string
          nullable: true
      required: [title]

    GeneratedImage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: integer
          nullable: true
        title:
          type: string
        imageUrl:
          type: string
        aspectRatio:
          type: string
        fileSize:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, imageUrl, aspectRatio, createdAt]

    HistoryResponse:
      type: object
      properties:
        images:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedImage"
        total:
          type: integer
      required: [images, total]

    ApiKeyResponse:
      type: object
      properties:
        apiKey:
          type: string
      required: [apiKey]

    StatsResponse:
      type: object
      properties:
        totalImages:
          type: integer
        imagesToday:
          type: integer
        totalTemplates:
          type: integer
        plan:
          type: string
        dailyLimit:
          type: integer
      required: [totalImages, imagesToday, totalTemplates, plan, dailyLimit]