# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * News Card Generator Pro API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    plan: zod.string(),
    apiKey: zod.string(),
    imagesToday: zod.number(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  plan: zod.string(),
  apiKey: zod.string(),
  imagesToday: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List user templates
 */
export const ListTemplatesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  slug: zod.string().nullish(),
  backgroundUrl: zod.string().nullish(),
  logoUrl: zod.string().nullish(),
  overlayUrl: zod.string().nullish(),
  elements: zod.string(),
  category: zod.string(),
  aspectRatio: zod.string(),
  bannerColor: zod.string(),
  bannerGradient: zod.string().nullish(),
  textColor: zod.string(),
  labelColor: zod.string(),
  font: zod.string(),
  fontSize: zod.number(),
  fontWeight: zod.number(),
  photoHeight: zod.number(),
  subtitle: zod.string().nullish(),
  label: zod.string().nullish(),
  logoText: zod.string().nullish(),
  logoPos: zod.string(),
  logoInvert: zod.boolean(),
  textShadow: zod.boolean(),
  isPublic: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListTemplatesResponse = zod.array(ListTemplatesResponseItem);

/**
 * @summary Create a template
 */
export const CreateTemplateBody = zod.object({
  name: zod.string(),
  slug: zod.string().nullish(),
  backgroundUrl: zod.string().nullish(),
  logoUrl: zod.string().nullish(),
  overlayUrl: zod.string().nullish(),
  elements: zod.string().optional(),
  category: zod.string().optional(),
  aspectRatio: zod.string().optional(),
  bannerColor: zod.string().optional(),
  bannerGradient: zod.string().nullish(),
  textColor: zod.string().optional(),
  labelColor: zod.string().optional(),
  font: zod.string().optional(),
  fontSize: zod.number().optional(),
  fontWeight: zod.number().optional(),
  photoHeight: zod.number().optional(),
  subtitle: zod.string().nullish(),
  label: zod.string().nullish(),
  logoText: zod.string().nullish(),
  logoPos: zod.string().optional(),
  logoInvert: zod.boolean().optional(),
  textShadow: zod.boolean().optional(),
  isPublic: zod.boolean().optional(),
});

/**
 * @summary Get a template
 */
export const GetTemplateParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTemplateResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  slug: zod.string().nullish(),
  backgroundUrl: zod.string().nullish(),
  logoUrl: zod.string().nullish(),
  overlayUrl: zod.string().nullish(),
  elements: zod.string(),
  category: zod.string(),
  aspectRatio: zod.string(),
  bannerColor: zod.string(),
  bannerGradient: zod.string().nullish(),
  textColor: zod.string(),
  labelColor: zod.string(),
  font: zod.string(),
  fontSize: zod.number(),
  fontWeight: zod.number(),
  photoHeight: zod.number(),
  subtitle: zod.string().nullish(),
  label: zod.string().nullish(),
  logoText: zod.string().nullish(),
  logoPos: zod.string(),
  logoInvert: zod.boolean(),
  textShadow: zod.boolean(),
  isPublic: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a template
 */
export const UpdateTemplateParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTemplateBody = zod.object({
  name: zod.string(),
  slug: zod.string().nullish(),
  backgroundUrl: zod.string().nullish(),
  logoUrl: zod.string().nullish(),
  overlayUrl: zod.string().nullish(),
  elements: zod.string().optional(),
  category: zod.string().optional(),
  aspectRatio: zod.string().optional(),
  bannerColor: zod.string().optional(),
  bannerGradient: zod.string().nullish(),
  textColor: zod.string().optional(),
  labelColor: zod.string().optional(),
  font: zod.string().optional(),
  fontSize: zod.number().optional(),
  fontWeight: zod.number().optional(),
  photoHeight: zod.number().optional(),
  subtitle: zod.string().nullish(),
  label: zod.string().nullish(),
  logoText: zod.string().nullish(),
  logoPos: zod.string().optional(),
  logoInvert: zod.boolean().optional(),
  textShadow: zod.boolean().optional(),
  isPublic: zod.boolean().optional(),
});

export const UpdateTemplateResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  slug: zod.string().nullish(),
  backgroundUrl: zod.string().nullish(),
  logoUrl: zod.string().nullish(),
  overlayUrl: zod.string().nullish(),
  elements: zod.string(),
  category: zod.string(),
  aspectRatio: zod.string(),
  bannerColor: zod.string(),
  bannerGradient: zod.string().nullish(),
  textColor: zod.string(),
  labelColor: zod.string(),
  font: zod.string(),
  fontSize: zod.number(),
  fontWeight: zod.number(),
  photoHeight: zod.number(),
  subtitle: zod.string().nullish(),
  label: zod.string().nullish(),
  logoText: zod.string().nullish(),
  logoPos: zod.string(),
  logoInvert: zod.boolean(),
  textShadow: zod.boolean(),
  isPublic: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a template
 */
export const DeleteTemplateParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Generate a news card image
 */
export const GenerateImageBody = zod.object({
  title: zod.string(),
  subtitle: zod.string().nullish(),
  label: zod.string().nullish(),
  templateId: zod.union([zod.string(), zod.number()]).nullish(),
  aspectRatio: zod.string().optional(),
  backgroundPhotoFilename: zod.string().nullish(),
  logoPhotoFilename: zod.string().nullish(),
});

/**
 * @summary Get generation history
 */
export const listHistoryQueryLimitDefault = 20;
export const listHistoryQueryOffsetDefault = 0;

export const ListHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(listHistoryQueryLimitDefault),
  offset: zod.coerce.number().default(listHistoryQueryOffsetDefault),
});

export const ListHistoryResponse = zod.object({
  images: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      templateId: zod.number().nullish(),
      title: zod.string(),
      imageUrl: zod.string(),
      aspectRatio: zod.string(),
      fileSize: zod.number().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Regenerate API key
 */
export const RegenerateApiKeyResponse = zod.object({
  apiKey: zod.string(),
});

/**
 * @summary Get dashboard stats
 */
export const GetStatsResponse = zod.object({
  totalImages: zod.number(),
  imagesToday: zod.number(),
  totalTemplates: zod.number(),
  plan: zod.string(),
  dailyLimit: zod.number(),
});