# lib/api-zod/src/generated/types/createTemplateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * News Card Generator Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTemplateBody {
  name: string;
  slug?: string | null;
  backgroundUrl?: string | null;
  logoUrl?: string | null;
  overlayUrl?: string | null;
  elements?: string;
  category?: string;
  aspectRatio?: string;
  bannerColor?: string;
  bannerGradient?: string | null;
  textColor?: string;
  labelColor?: string;
  font?: string;
  fontSize?: number;
  fontWeight?: number;
  photoHeight?: number;
  subtitle?: string | null;
  label?: string | null;
  logoText?: string | null;
  logoPos?: string;
  logoInvert?: boolean;
  textShadow?: boolean;
  isPublic?: boolean;
}