# lib/api-zod/src/generated/types/generateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * News Card Generator Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateBody {
  title: string;
  subtitle?: string | null;
  label?: string | null;
  templateId?: string | number | null;
  aspectRatio?: string;
  backgroundPhotoFilename?: string | null;
  logoPhotoFilename?: string | null;
}