# lib/api-zod/src/generated/types/generatedImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * News Card Generator Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedImage {
  id: number;
  userId: number;
  templateId?: number | null;
  title: string;
  imageUrl: string;
  aspectRatio: string;
  fileSize?: number | null;
  createdAt: Date;
}